requirements-base.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.0.1
motor==3.3.2
pymongo==4.6.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.1.0 